fix(03-02): use u32 for stash index (specta compatibility)

+++ src-tauri/src/git/error.rs
@@ -71,7 +71,7 @@ pub enum GitError {
 
     // Stash errors
     #[error("Stash not found at index {0}")]
-    StashNotFound(usize),
+    StashNotFound(u32),
 
     #[error("No local changes to stash")]
     NothingToStash,
